methods update & delete

+++ src/main/java/com/umanizales/watchman_app/infrastructure/adapters/PostgresWatchmanRepository.java
@@ -2,6 +2,7 @@ package com.umanizales.watchman_app.infrastructure.adapters;
 
 import com.umanizales.watchman_app.aplication.WatchmanAble;
 import com.umanizales.watchman_app.domain.WatchmanDTO;
+import com.umanizales.watchman_app.exception.WatchmanAppException;
 import com.umanizales.watchman_app.infrastructure.repositories.entity.WatchmanEntity;
 import com.umanizales.watchman_app.infrastructure.repositories.repository.WatchmanRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +14,7 @@ import java.util.Collections;
 import java.util.List;
 
 @Qualifier("PostgresWatchmanRepository")
-@Repository("WatchmanPersistance")
+@Repository("WatchmanPersistence")
 public class PostgresWatchmanRepository implements WatchmanAble {
 
     //inyeccion de dependencias
@@ -27,14 +28,31 @@ public class PostgresWatchmanRepository implements WatchmanAble {
     }
 
     @Override
-    public WatchmanDTO update(WatchmanDTO watchmanDTO) {
-        return watchmanRepository.save(new WatchmanEntity(watchmanDTO)).toWatchmanDTO();
+    public WatchmanDTO update(int code, WatchmanDTO watchmanDTO) throws WatchmanAppException {
+        if (watchmanRepository.existsById(code)){
+             return watchmanRepository.save(new WatchmanEntity(watchmanDTO)).toWatchmanDTO();
+        }
+        else {
+            throw new WatchmanAppException("El codigo a editar no existe: " + code);
+        }
     }
 
     @Override
-    public boolean delete(int code) {
-        watchmanRepository.deleteAllById(Collections.singleton(code));
-        return true;
+    public boolean delete(int code) throws WatchmanAppException {
+
+        if(watchmanRepository.existsById(code)){
+            try {
+                watchmanRepository.deleteAllById(Collections.singleton(code));
+                return true;
+            }
+            catch (Exception e){
+                throw new WatchmanAppException("Este watchman no se puede eliminar, tiene relaciones: " + code);
+            }
+        }
+        else {
+            throw new WatchmanAppException("El codigo a borrar no existe: " + code);
+        }
+
     }
 
     @Override
